requirements.txt
aiohttp
async-timeout
attrs
chardet
idna
multidict
slack-sdk
typing-extensions
yarl
psycopg2-binary
psutil